+++ worker/src/utils/validation.ts
